pyproject.toml
[project]
name = "ai-medical-imaging-open"
version = "0.1.0"
description = "Reproducible Open-MELON melanoma classification experiments"
readme = "REPO_RULES.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "datasets>=4.1.0",
  "numpy>=2.0.0",
  "pillow>=11.0.0",
  "protobuf>=7.34.1",
  "pyyaml>=6.0.0",
  "scikit-learn>=1.5.0",
  "sentencepiece>=0.2.1",
]

[project.optional-dependencies]
captioning = [
  "google-genai>=1.0.0",
  "ollama>=0.4.0",
]
dermatollama = [
  "accelerate>=1.8.0",
  "peft>=0.16.0",
  "torch>=2.6.0",
  "transformers>=4.45.0",
]
embeddings = [
  "open-clip-torch>=2.26.0",
  "torch>=2.6.0",
  "transformers>=4.45.0",
]
dev = [
  "accelerate>=1.8.0",
  "google-genai>=1.0.0",
  "ollama>=0.4.0",
  "open-clip-torch>=2.26.0",
  "peft>=0.16.0",
  "torch>=2.6.0",
  "transformers>=4.45.0",
]

[tool.uv]
default-groups = []
